# ergon-dashboard/src/generated/rest/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Ergon",
        "description": "Ergon experiment orchestration API",
        "version": "0.1.0"
    },
    "paths": {
        "/runs/{run_id}": {
            "get": {
                "tags": [
                    "runs"
                ],
                "summary": "Get Run",
                "description": "Get a persisted run-detail snapshot suitable for frontend hydration.",
                "operationId": "get_run_runs__run_id__get",
                "parameters": [
                    {
                        "name": "run_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Run Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RunSnapshotDto"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/cohorts": {
            "get": {
                "tags": [
                    "cohorts"
                ],
                "summary": "List Cohorts",
                "description": "List all experiment cohorts.",
                "operationId": "list_cohorts_cohorts_get",
                "parameters": [
                    {
                        "name": "include_archived",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "default": false,
                            "title": "Include Archived"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CohortSummaryDto"
                                    },
                                    "title": "Response List Cohorts Cohorts Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/cohorts/{cohort_id}": {
            "get": {
                "tags": [
                    "cohorts"
                ],
                "summary": "Get Cohort",
                "description": "Get one cohort detail payload.",
                "operationId": "get_cohort_cohorts__cohort_id__get",
                "parameters": [
                    {
                        "name": "cohort_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Cohort Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CohortDetailDto"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "cohorts"
                ],
                "summary": "Update Cohort",
                "description": "Update one cohort's operator-managed fields.",
                "operationId": "update_cohort_cohorts__cohort_id__patch",
                "parameters": [
                    {
                        "name": "cohort_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Cohort Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCohortRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CohortSummaryDto"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "CohortDetailDto": {
                "properties": {
                    "summary": {
                        "$ref": "#/components/schemas/CohortSummaryDto"
                    },
                    "runs": {
                        "items": {
                            "$ref": "#/components/schemas/CohortRunRowDto"
                        },
                        "type": "array",
                        "title": "Runs"
                    }
                },
                "type": "object",
                "required": [
                    "summary"
                ],
                "title": "CohortDetailDto",
                "description": "Full payload for a single cohort detail page."
            },
...
            "CohortSummaryDto": {
                "properties": {
                    "cohort_id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Cohort Id"
                    },
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "description": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Description"
                    },
                    "created_by": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Created By"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    },
                    "status": {
                        "type": "string",
                        "title": "Status"
                    },
                    "total_runs": {
                        "type": "integer",
                        "title": "Total Runs",
                        "default": 0
                    },
                    "status_counts": {
                        "$ref": "#/components/schemas/CohortStatusCountsDto"
                    },
                    "average_score": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Average Score"
                    },
                    "best_score": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Best Score"
                    },
                    "worst_score": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Worst Score"
                    },
                    "average_duration_ms": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Average Duration Ms"
                    },
                    "failure_rate": {
                        "type": "number",
                        "title": "Failure Rate",
                        "default": 0.0
                    },
                    "stats_updated_at": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Stats Updated At"
                    }
                },
                "type": "object",
                "required": [
                    "cohort_id",
                    "name",
                    "created_at",
                    "status"
                ],
                "title": "CohortSummaryDto",
                "description": "Summary row for cohort list and live updates."
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "RunActionDto": {
                "properties": {
                    "id": {
                        "type": "string",
                        "title": "Id"
                    },
                    "taskId": {
                        "type": "string",
                        "title": "Taskid"
                    },
                    "workerId": {
                        "type": "string",
                        "title": "Workerid"
                    },
                    "workerName": {
                        "type": "string",
                        "title": "Workername"
                    },
                    "type": {
                        "type": "string",
                        "title": "Type"
                    },
                    "input": {
                        "type": "string",
                        "title": "Input"
                    },
                    "output": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Output"
                    },
                    "status": {
                        "type": "string",
                        "title": "Status"
                    },
                    "success": {
                        "type": "boolean",
                        "title": "Success"
                    },
                    "startedAt": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Startedat"
                    },
                    "completedAt": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Completedat"
                    },
                    "durationMs": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Durationms"
                    },
                    "error": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Error"
                    }
                },
                "additionalProperties": false,
                "type": "object",
                "required": [
                    "id",
                    "taskId",
                    "workerId",
                    "workerName",
                    "type",
                    "input",
                    "status",
                    "success"
                ],
                "title": "RunActionDto"
            },
...
            "RunResourceDto": {
                "properties": {
                    "id": {
                        "type": "string",
                        "title": "Id"
                    },
                    "taskId": {
                        "type": "string",
                        "title": "Taskid"
                    },
                    "taskExecutionId": {
                        "type": "string",
                        "title": "Taskexecutionid"
                    },
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "mimeType": {
                        "type": "string",
                        "title": "Mimetype"
                    },
                    "filePath": {
                        "type": "string",
                        "title": "Filepath"
                    },
                    "sizeBytes": {
                        "type": "integer",
                        "title": "Sizebytes"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Createdat"
                    }
                },
                "additionalProperties": false,
                "type": "object",
                "required": [
                    "id",
                    "taskId",
                    "taskExecutionId",
                    "name",
                    "mimeType",
                    "filePath",
                    "sizeBytes",
                    "createdAt"
                ],
                "title": "RunResourceDto"
            },
...
            "RunSnapshotDto": {
                "properties": {
                    "id": {
                        "type": "string",
                        "title": "Id"
                    },
                    "experimentId": {
                        "type": "string",
                        "title": "Experimentid"
                    },
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "status": {
                        "type": "string",
                        "title": "Status"
                    },
                    "tasks": {
                        "additionalProperties": {
                            "$ref": "#/components/schemas/RunTaskDto"
                        },
                        "type": "object",
                        "title": "Tasks"
                    },
                    "rootTaskId": {
                        "type": "string",
                        "title": "Roottaskid",
                        "default": ""
                    },
                    "actionsByTask": {
                        "additionalProperties": {
                            "items": {
                                "$ref": "#/components/schemas/RunActionDto"
                            },
                            "type": "array"
                        },
                        "type": "object",
                        "title": "Actionsbytask"
                    },
                    "resourcesByTask": {
                        "additionalProperties": {
                            "items": {
                                "$ref": "#/components/schemas/RunResourceDto"
                            },
                            "type": "array"
                        },
                        "type": "object",
                        "title": "Resourcesbytask"
                    },
                    "executionsByTask": {
                        "additionalProperties": {
                            "items": {
                                "$ref": "#/components/schemas/RunExecutionAttemptDto"
                            },
                            "type": "array"
                        },
                        "type": "object",
                        "title": "Executionsbytask"
                    },
                    "evaluationsByTask": {
                        "additionalProperties": {
                            "$ref": "#/components/schemas/RunTaskEvaluationDto"
                        },
                        "type": "object",
                        "title": "Evaluationsbytask"
                    },
                    "sandboxesByTask": {
                        "additionalProperties": {
                            "$ref": "#/components/schemas/RunSandboxDto"
                        },
                        "type": "object",
                        "title": "Sandboxesbytask"
                    },
                    "threads": {
                        "items": {
                            "$ref": "#/components/schemas/RunCommunicationThreadDto"
                        },
                        "type": "array",
                        "title": "Threads"
                    },
                    "startedAt": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Startedat"
                    },
                    "completedAt": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Completedat"
                    },
                    "durationSeconds": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Durationseconds"
                    },
                    "totalTasks": {
                        "type": "integer",
                        "title": "Totaltasks",
                        "default": 0
                    },
                    "totalLeafTasks": {
                        "type": "integer",
                        "title": "Totalleaftasks",
                        "default": 0
                    },
                    "completedTasks": {
                        "type": "integer",
                        "title": "Completedtasks",
                        "default": 0
                    },
                    "failedTasks": {
                        "type": "integer",
                        "title": "Failedtasks",
                        "default": 0
                    },
                    "runningTasks": {
                        "type": "integer",
                        "title": "Runningtasks",
                        "default": 0
                    },
                    "finalScore": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Finalscore"
                    },
                    "error": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Error"
                    }
                },
                "additionalProperties": false,
                "type": "object",
                "required": [
                    "id",
                    "experimentId",
                    "name",
                    "status"
                ],
                "title": "RunSnapshotDto"
            },
            "RunTaskDto": {
                "properties": {
                    "id": {
                        "type": "string",
                        "title": "Id"
                    },
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "description": {
                        "type": "string",
                        "title": "Description"
                    },
                    "status": {
                        "type": "string",
                        "title": "Status"
                    },
                    "parentId": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Parentid"
                    },
                    "childIds": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Childids"
                    },
                    "dependsOnIds": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Dependsonids"
                    },
                    "isLeaf": {
                        "type": "boolean",
                        "title": "Isleaf"
                    },
                    "level": {
                        "type": "integer",
                        "title": "Level"
                    },
                    "assignedWorkerId": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Assignedworkerid"
                    },
                    "assignedWorkerName": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Assignedworkername"
                    },
                    "startedAt": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Startedat"
                    },
                    "completedAt": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Completedat"
                    }
                },
                "additionalProperties": false,
                "type": "object",
                "required": [
                    "id",
                    "name",
                    "description",
                    "status",
                    "isLeaf",
                    "level"
                ],
                "title": "RunTaskDto"
            },
...
            "UpdateCohortRequest": {
                "properties": {
                    "status": {
                        "$ref": "#/components/schemas/ExperimentCohortStatus"
                    }
                },
                "type": "object",
                "required": [
                    "status"
                ],
                "title": "UpdateCohortRequest",
                "description": "Mutable cohort fields exposed through the operator API."
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    },
                    "input": {
                        "title": "Input"
                    },
                    "ctx": {
                        "type": "object",
                        "title": "Context"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}